requirements.txt.txt
streamlit
tensorflow-cpu
Pillow
numpy
pandas
plotly